chore(CoreModule): reorganize file structure
