+++ src/app/core/index.ts
@@ -1,0 +1,4 @@
+export { Selector } from './interfaces';
+export { ApiService } from './services/api';
+export { MediaQueryResults, MediaQueryService } from './services/media-query';
+export { CoreModule } from './core-module';
