+++ src/app/core/interfaces/index.ts
@@ -1,0 +1,1 @@
+export { Selector } from './selector';
