+++ src/app/core/interfaces/selector.ts
@@ -1,0 +1,4 @@
+import { Observable } from 'rxjs/Observable';
+
+
+export type Selector<T,V> = (observable$: Observable<T>) => Observable<V>;
