+++ src/app/core/services/api/interfaces.ts
