+++ src/app/core/services/media-query/interfaces.ts
