+++ src/app/core/services/media-query/media-query-list-stub.ts
