--- src/core/interfaces.ts
@@ -1,6 +1,0 @@
-import { Observable } from 'rxjs/Observable';
-
-
-export interface Selector<T,V> {
-  (observable$: Observable<T>): Observable<V>;
-}
