+++ src/tracklists/components/tracklist.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
-import { MediaQueryService } from 'src/core';
+import { MediaQueryService } from 'src/app/core';
 import { PlayerService } from 'src/player';
 import { TracklistService } from '../tracklist-service';
 import { TracklistScrollService } from './tracklist-scroll-service';
