+++ src/tracklists/reducers/selectors.ts
@@ -7,7 +7,7 @@ import 'rxjs/add/operator/withLatestFrom';
 import { List } from 'immutable';
 import { AppState } from 'src/app';
 import { TRACKS_PER_PAGE } from 'src/constants';
-import { Selector } from 'src/core';
+import { Selector } from 'src/app/core';
 import { Tracklist } from '../models/tracklist';
 import { Track } from '../models/track';
 import { TracklistsState } from './tracklists-reducer';
